# Conflicts:
#	src/main/kotlin/cn/yiiguxing/plugin/translate/Settings.kt
#	src/main/kotlin/cn/yiiguxing/plugin/translate/ui/TranslationDialog.kt
#	src/main/resources/messages/TranslationBundle.properties
#	src/main/resources/messages/TranslationBundle_ja.properties
#	src/main/resources/messages/TranslationBundle_ko.properties
#	src/main/resources/messages/TranslationBundle_zh_CN.properties

+++ src/main/kotlin/cn/yiiguxing/plugin/translate/trans/openai/chat/ChatCompletion.kt
@@ -16,7 +16,7 @@ data class ChatCompletion internal constructor(
     /**
      * The creation time in epoch milliseconds.
      */
-    @SerializedName("created") val created: Int,
+    @SerializedName("created") val created: Long,
 
     /**
      * The model used.
